Move to senaite.core.api.workflow
